refactor: consolidate retry/sleep/error utilities from providers into shared utils
Extract duplicated sleep(), isRetryableError(), isTransientNetworkError(),
and extractRetryDelay/extractRetryAfterMs into a single shared module at
packages/pi-ai/src/utils/retry-utils.ts.

The unified extractRetryDelayMs() handles all header formats from both
Google (x-ratelimit-reset, x-ratelimit-reset-after) and Anthropic
(x-ratelimit-reset-requests, x-ratelimit-reset-tokens), plus body text
parsing patterns from the Gemini provider. The unified isRetryableError()
merges regex patterns from both Google and OpenAI providers.

Removes ~213 lines of duplication across 3 provider files.
