+++ packages/pi-ai/src/providers/anthropic.ts
@@ -30,6 +30,7 @@ import type {
 } from "../types.js";
 import { AssistantMessageEventStream } from "../utils/event-stream.js";
 import { parseStreamingJson } from "../utils/json-parse.js";
+import { extractRetryDelayMs, isTransientNetworkError } from "../utils/retry-utils.js";
 import { sanitizeSurrogates } from "../utils/sanitize-unicode.js";
 
 import { buildCopilotDynamicHeaders, hasCopilotVisionInput } from "./github-copilot-headers.js";
@@ -201,65 +202,6 @@ function mergeHeaders(...headerSources: (Record<string, string> | undefined)[]):
 		}
 	}
 	return merged;
-}
-
-/**
- * Detect transient network errors that are likely to succeed on retry.
- * Covers WebSocket disconnects (Tailscale, VPN), TCP resets, and DNS failures.
- */
-function isTransientNetworkError(error: unknown): boolean {
-  if (!(error instanceof Error)) return false;
-  const msg = error.message.toLowerCase();
-  const code = (error as NodeJS.ErrnoException).code;
-  return (
-    code === 'ECONNRESET' ||
-    code === 'EPIPE' ||
-    code === 'ETIMEDOUT' ||
-    code === 'ENOTFOUND' ||
-    code === 'EAI_AGAIN' ||
-    msg.includes('connector_closed') ||
-    msg.includes('socket hang up') ||
-    msg.includes('network') ||
-    msg.includes('connection') && msg.includes('closed') ||
-    msg.includes('fetch failed')
-  );
-}
-
-/**
- * Extract retry delay from Anthropic error response headers (in milliseconds).
- * Checks: retry-after (seconds or RFC date), x-ratelimit-reset-requests, x-ratelimit-reset-tokens.
- * Returns undefined if no valid delay is found or if the delay is in the past.
- */
-export function extractRetryAfterMs(headers: Headers | { get(name: string): string | null }, errorText = ""): number | undefined {
-	const normalizeDelay = (ms: number): number | undefined => (ms > 0 ? Math.ceil(ms + 1000) : undefined);
-
-	const retryAfter = headers.get("retry-after");
-	if (retryAfter) {
-		const seconds = Number(retryAfter);
-		if (Number.isFinite(seconds)) {
-			const delay = normalizeDelay(seconds * 1000);
-			if (delay !== undefined) return delay;
-		}
-		const asDate = new Date(retryAfter).getTime();
-		if (!Number.isNaN(asDate)) {
-			const delay = normalizeDelay(asDate - Date.now());
-			if (delay !== undefined) return delay;
-		}
-	}
-
-	// x-ratelimit-reset-requests / x-ratelimit-reset-tokens are Unix timestamps (seconds)
-	for (const header of ["x-ratelimit-reset-requests", "x-ratelimit-reset-tokens"]) {
-		const value = headers.get(header);
-		if (value) {
-			const resetSeconds = Number(value);
-			if (Number.isFinite(resetSeconds)) {
-				const delay = normalizeDelay(resetSeconds * 1000 - Date.now());
-				if (delay !== undefined) return delay;
-			}
-		}
-	}
-
-	return undefined;
 }
 
 export const streamAnthropic: StreamFunction<"anthropic-messages", AnthropicOptions> = (
@@ -514,7 +456,7 @@ export const streamAnthropic: StreamFunction<"anthropic-messages", AnthropicOpti
 			}
 			const AnthropicSdk = _AnthropicClass;
 			if (AnthropicSdk && error instanceof AnthropicSdk.APIError && error.headers) {
-				const retryAfterMs = extractRetryAfterMs(error.headers, error.message);
+				const retryAfterMs = extractRetryDelayMs(error.headers, error.message);
 				if (retryAfterMs !== undefined) {
 					output.retryAfterMs = retryAfterMs;
 				}
